Refactor kfp compiler module
