+++ compilers/Makefile
@@ -1,7 +1,7 @@
-include ../../docker-targets.mk
-include ../../help.mk
+include ../docker-targets.mk
+include ../help.mk
 
-IMG := kfp-operator-argo-kfp-compiler
+IMG := kfp-operator-tfx-compiler
 
 all: build
 
@@ -29,7 +29,7 @@ integration-test: docker-build penguin-38 penguin-39 penguin-310 ## Run the inte
 	$(eval TMP := $(shell mktemp -d))
 	echo '{"rootLocation": "pipeline_root", "servingLocation": "serving", "name": "test", "image": "test-pipeline", "tfxComponents": "pipeline.create_components"}' > $(TMP)/pipeline_config.json
 	docker run -v $(TMP):/shared kfp-operator-argo-kfp-compiler /shared
-	docker run -v $(TMP):/shared --entrypoint /shared/compile.sh penguin:3.10 --output_file /tmp/pipeline.yaml --pipeline_config /shared/pipeline_config.json  --execution_mode v1
+
 	docker run -v $(TMP):/shared --entrypoint /shared/compile.sh penguin:3.8 --output_file /tmp/pipeline.yaml --pipeline_config /shared/pipeline_config.json --execution_mode v2
 	docker run -v $(TMP):/shared --entrypoint /shared/compile.sh penguin:3.9 --output_file /tmp/pipeline.yaml --pipeline_config /shared/pipeline_config.json --execution_mode v2
 	docker run -v $(TMP):/shared --entrypoint /shared/compile.sh penguin:3.10 --output_file /tmp/pipeline.yaml --pipeline_config /shared/pipeline_config.json  --execution_mode v2
@@ -47,5 +47,4 @@ WHEEL_VERSION=$(shell poetry version | cut -d ' ' -f 2)
 DOCKER_BUILD_EXTRA_PARAMS=--build-arg WHEEL_VERSION=${WHEEL_VERSION}
 
 docker-push-all: ## Push all compiler docker images
-	$(MAKE) docker-push IMG=kfp-operator-argo-kfp-compiler
 	$(MAKE) docker-push IMG=kfp-operator-tfx-compiler
