+++ compilers/resources/compile.sh
@@ -4,4 +4,4 @@ PYTHON_VERSION=$(python3 -V | grep -Eo '[0-9]+\.[0-9]+')
 
 export PYTHONPATH="$PYTHONPATH:$(dirname $0)/py$PYTHON_VERSION"
 
-python3 -m 'kfp_compiler' "$@"
+python3 -m 'tfx_compiler' "$@"
