+++ compilers/tfx/TFX-Dockerfile
@@ -25,9 +25,9 @@ WORKDIR /
 COPY --from=PYTHON38 /kfp-compiler kfp-compiler/py3.8
 COPY --from=PYTHON39 /kfp-compiler kfp-compiler/py3.9
 COPY --from=PYTHON310 /kfp-compiler kfp-compiler/py3.10
-ADD resources/compile.sh kfp-compiler/compile.sh
+ADD ../resources/compile.sh tfx-compiler/compile.sh
 
-ADD resources/entrypoint.sh entrypoint.sh
+ADD ../resources/entrypoint.sh entrypoint.sh
 
 USER 65534:65534
 ENTRYPOINT ["/entrypoint.sh"]
