+++ compilers/tfx/integration/Dockerfile
