+++ config/crd/bases/config.kubeflow.org_kfpcontrollerconfigs.yaml
@@ -7698,6 +7698,10 @@ spec:
                   servicePort:
                     type: integer
                 type: object
+              frameworks:
+                additionalProperties:
+                  type: string
+                type: object
               multiversion:
                 type: boolean
               pipelineFrameworkImages:
